requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
termcolor==2.4.0
python-dateutil==2.8.2
uuid==1.30 